[refactor] (session) add session package
